requirements.txt
streamlit>=1.32.0
tensorflow-cpu>=2.16.1
numpy>=1.26.0
matplotlib>=3.8.2
scipy>=1.12.0
seaborn>=0.13.1
